restructure project into modules
